+++ src/lower/merge_lattice_old.cpp
@@ -85,7 +85,7 @@ MergeLattice MergeLattice::make(Forall forall,
     }
 
     Iterator getIterator(Access access) {
-      size_t loc = util::locate(access.getIndexVars(),i) + 1;
+      int loc = (int)util::locate(access.getIndexVars(),i) + 1;
       taco_iassert(util::contains(iterators, ModeAccess(access,loc)));
       return iterators.at(ModeAccess(access,loc));
     }
