+++ src/lower/mode_access.cpp
@@ -2,7 +2,7 @@
 
 namespace taco {
 
-ModeAccess::ModeAccess(Access access, size_t mode) : access(access), mode(mode){
+ModeAccess::ModeAccess(Access access, int mode) : access(access), mode(mode){
 }
 
 Access ModeAccess::getAccess() const {
